extract broadcasted events tests
